clearcut_detection_backend/test/requirements.txt
affine==2.3.0
argcomplete==1.11.1
attrs==19.3.0
boto==2.49.0
boto3==1.14.11
botocore==1.17.11
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
docutils==0.15.2
fasteners==0.15
Fiona==1.8.13.post1
gcs-oauth2-boto-plugin==2.6
geopandas==0.5.1
google-api-core==1.14.3
google-api-python-client==1.7.4
google-apitools==0.5.31
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-reauth==0.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gsutil==4.47
httplib2==0.18.1
idna==2.9
imageio==2.5.0
importlib-metadata==1.6.1
jmespath==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.2
mercantile==1.1.5
mock==2.0.0
monotonic==1.5
munch==2.5.0
networkx==2.4
numpy==1.16.4
oauth2client==4.1.3
opencv-python==4.2.0.34
pandas==0.24.2
pbr==5.4.5
Pillow==6.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.2.1
python-dateutil==2.8.1
pytz==2020.1
pyu2f==0.1.4
PyWavelets==1.1.1
PyYAML==5.1.1
rasterio==1.1.2
requests==2.24.0
retry-decorator==1.1.1
rio-mbtiles==1.4.2
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.0
sentinelhub==2.5.3
Shapely==1.6.4.post2
six==1.12.0
snuggs==1.4.7
SocksiPy-branch==1.1
tifffile==2020.6.3
tqdm==4.19.9
uritemplate==3.0.1
urllib3==1.25.9
utm==0.5.0
zipp==3.1.0

catalyst==19.5
segmentation-models-pytorch==0.1.0
torch==1.4.0
torchvision==0.5.0
